#81 StartViewController에서 Realm 관련 로직을 extension으로 분리
